+++ lib/builder.dart
@@ -40,7 +40,8 @@ class Builder extends Command {
   ArgParser get argParser {
     environment = Environment.fromArgResults(globalResults);
     return ArgParser()
-      ..addFlag("publish", abbr: "p", defaultsTo: false, help: "Distribute Android")
+      ..addFlag("publish",
+          abbr: "p", defaultsTo: false, help: "Distribute Android")
       ..addFlag(
         "android",
         defaultsTo: environment.isAndroidBuild,
@@ -51,12 +52,14 @@ class Builder extends Command {
         defaultsTo: Platform.isMacOS ? environment.isIOSBuild : false,
         help: "Build iOS (Default value follows the config file)",
       )
-      ..addOption("android_args", defaultsTo: "", help: "Arguments for Android build.")
+      ..addOption("android_args",
+          defaultsTo: "", help: "Arguments for Android build.")
       ..addOption("ios_args", defaultsTo: "", help: "Arguments for iOS build.")
       ..addOption(
         'custom_args',
         defaultsTo: "",
-        help: "Custom arguments key:args,key:args, it will executed as `flutter build <args>`",
+        help:
+            "Custom arguments key:args,key:args, it will executed as `flutter build <args>`",
         valueHelp: "macos:macos,windows:windows,ios:ipa,android_apk:apk",
       );
   }
@@ -84,7 +87,8 @@ class Builder extends Command {
     final customBuildArgs = <String, List<String>>{};
 
     if (!await environment.initialized) {
-      ColorizeLogger.logError("Please run distribute init first ${await environment.initialized}");
+      ColorizeLogger.logError(
+          "Please run distribute init first ${await environment.initialized}");
       exit(1);
     }
 
@@ -106,7 +110,11 @@ class Builder extends Command {
     }
     if (iosArgs.isNotEmpty) customBuildArgs['ios'] = iosArgs.split(',');
 
-    return build(androidArgs: customBuildArgs['android'] ?? [], iosArgs: customBuildArgs['ios'] ?? [], customBuildArgs: customBuildArgs).then((value) {
+    return build(
+            androidArgs: customBuildArgs['android'] ?? [],
+            iosArgs: customBuildArgs['ios'] ?? [],
+            customBuildArgs: customBuildArgs)
+        .then((value) {
       if (value != 0) {
         ColorizeLogger.logError('Build failed.');
         exit(1);
@@ -171,12 +179,20 @@ class Builder extends Command {
     if (publish) {
       List<(String, int)> distributionResults = [];
       for (var buildResult in buildResults) {
-        if (buildAndroid && buildResult.$1 == "android" && buildResult.$2 == 0) {
+        if (buildAndroid &&
+            buildResult.$1 == "android" &&
+            buildResult.$2 == 0) {
           await publisher.buildAndroidDocs();
-          distributionResults.add(await publisher.distributeAndroid().then((value) => (buildResult.$1, value)).catchError((e) => (buildResult.$1, 1)));
+          distributionResults.add(await publisher
+              .distributeAndroid()
+              .then((value) => (buildResult.$1, value))
+              .catchError((e) => (buildResult.$1, 1)));
         }
         if (buildIOS && buildResult.$1 == "ios" && buildResult.$2 == 0) {
-          distributionResults.add(await publisher.distributeIOS().then((value) => (buildResult.$1, value)).catchError((e) => (buildResult.$1, 1)));
+          distributionResults.add(await publisher
+              .distributeIOS()
+              .then((value) => (buildResult.$1, value))
+              .catchError((e) => (buildResult.$1, 1)));
         }
       }
 
@@ -186,7 +202,8 @@ class Builder extends Command {
             ColorizeLogger.logError('Distribution failed for ${result.$1}.');
             exitCode = 1;
           } else {
-            ColorizeLogger.logSuccess('Distribution completed successfully for ${result.$1}.');
+            ColorizeLogger.logSuccess(
+                'Distribution completed successfully for ${result.$1}.');
           }
         }
       } else {
@@ -201,7 +218,8 @@ class Builder extends Command {
   Future<int> _buildAndroid({final List<String>? args}) async {
     if (buildAndroid) {
       ColorizeLogger.logDebug('Start android build...');
-      final process = await Process.start('flutter', ['build', 'aab', if (args != null) ...args]);
+      final process = await Process.start(
+          'flutter', ['build', 'aab', if (args != null) ...args]);
       if (environment.isVerbose) {
         process.stdout.transform(utf8.decoder).listen((data) {
           if (data.trim().isNotEmpty) ColorizeLogger.logDebug(data);
@@ -221,7 +239,8 @@ class Builder extends Command {
   /// Builds the iOS app with the specified [args].
   Future<int> _buildIOS({final List<String>? args}) async {
     if (buildIOS) {
-      final process = await Process.start('flutter', ['build', 'ipa', if (args != null) ...args]);
+      final process = await Process.start(
+          'flutter', ['build', 'ipa', if (args != null) ...args]);
       ColorizeLogger.logDebug('Start ios build...');
       if (environment.isVerbose) {
         process.stdout.transform(utf8.decoder).listen((data) {
@@ -271,7 +290,8 @@ class Builder extends Command {
       if (distributionDirList.isEmpty) {
         Directory outputDir = Directory("build/app/outputs/bundle");
         if (!await outputDir.exists()) {
-          ColorizeLogger.logError('No appbundle found in build/app/outputs/bundle');
+          ColorizeLogger.logError(
+              'No appbundle found in build/app/outputs/bundle');
           return 1;
         }
         final outputDirList = await outputDir.list().toList();
@@ -279,13 +299,17 @@ class Builder extends Command {
         for (var item in outputDirList) {
           if (item is Directory) {
             final files = await item.list().toList();
-            final index = files.indexWhere((item) => item.path.endsWith(".aab"));
+            final index =
+                files.indexWhere((item) => item.path.endsWith(".aab"));
             if (index > -1) {
               final appbundle = files[index];
               if (appbundle is File) {
-                await appbundle.copy("distribution/android/output/${appbundle.path.split("/").last}");
-                appbundles.add(File("distribution/android/output/${appbundle.path.split("/").last}"));
-                ColorizeLogger.logDebug("${appbundles.length} copied to distribution/android/output");
+                await appbundle.copy(
+                    "distribution/android/output/${appbundle.path.split("/").last}");
+                appbundles.add(File(
+                    "distribution/android/output/${appbundle.path.split("/").last}"));
+                ColorizeLogger.logDebug(
+                    "${appbundles.length} copied to distribution/android/output");
               }
             }
           }
@@ -319,13 +343,17 @@ class Builder extends Command {
         for (var item in outputDirList) {
           if (item is Directory) {
             final files = await item.list().toList();
-            final index = files.indexWhere((item) => item.path.endsWith(".ipa"));
+            final index =
+                files.indexWhere((item) => item.path.endsWith(".ipa"));
             if (index > -1) {
               final ipa = files[index];
               if (ipa is File) {
-                await ipa.copy("distribution/ios/output/${ipa.path.split("/").last}");
-                ipas.add(File("distribution/ios/output/${ipa.path.split("/").last}"));
-                ColorizeLogger.logDebug("${ipas.length} copied to distribution/ios/output");
+                await ipa.copy(
+                    "distribution/ios/output/${ipa.path.split("/").last}");
+                ipas.add(File(
+                    "distribution/ios/output/${ipa.path.split("/").last}"));
+                ColorizeLogger.logDebug(
+                    "${ipas.length} copied to distribution/ios/output");
               }
             }
           }